+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The following changes have been made for each official Clockwork build.
 * Removed HL2RP scanner ranking reference
 * Fixed custom models being overridden by rank models
 * Fixed rank model reversion to faction models
+* Added ITEM.noPlacementUse and ITEM.noInventoryUse fields to determine whether or not an item can be used as a placed entity, or in the inventory, respectively.
 
 0.99
 ------